Fixed a repo pruning issue for the llvm.global_ctors global variable.
If Module (A.cpp) contains llvm.global_ctors but it is removed after global
variable optimizer, it will not be stored in the repository. When we compile
A.cpp again, llvm.global_ctors will not be pruned. During the second build,
the _GLOBAL__sub_I_A.cpp will be added into the .text.startup section. The
_GLOBAL__sub_I_A.cpp calls the __cxx_global_var_init function. The
__cxx_global_var_init function then called a removed function, which is caused
a segmentation fault.

To solve the pruning issue, a fake fragment for the removed llvm.global_ctors
has been added the repository.