Accumulate the dependents' hash to this fragment.
If a fragment has dependents and the dependents have an internal
linkage type, accumulate the dependents' hash to this fragment.

If multiple global objects have the same fragment in the same modular, multiple compilation members with the same digest value have been added into the compilation. When this fragment digest is updated, the digestes of all the corresponding compilation members are needed to be updated. In the previous fix, only one corresponding compilation member's digest has been updated. This commit updated all comrresponding compilation members.

Fix for <https://github.com/SNSystems/llvm-project-prepo/issues/17>.